// src/main/webapp/src/app/auto-parking-ts-api/model/resultGitVersion.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { GitVersion } from './gitVersion';


export interface ResultGitVersion { 
    error?: number;
    message?: string;
    result?: GitVersion;
}