// src/main/webapp/src/app/auto-parking-ts-api/model/calculateRequest.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CarParkLocation } from './carParkLocation';


export interface CalculateRequest { 
    commandList?: Array<CalculateRequest.CommandListEnum>;
    initialCarParkLocation?: CarParkLocation;
    maxX?: number;
    maxY?: number;
}
export namespace CalculateRequest {
    export type CommandListEnum = 'L' | 'R' | 'F';
    export const CommandListEnum = {
        L: 'L' as CommandListEnum,
        R: 'R' as CommandListEnum,
        F: 'F' as CommandListEnum
    };
}