// src/main/webapp/src/app/auto-parking-ts-api/model/carParkLocation.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface CarParkLocation { 
    headingStatus?: CarParkLocation.HeadingStatusEnum;
    x?: number;
    y?: number;
}
export namespace CarParkLocation {
    export type HeadingStatusEnum = 'North' | 'South' | 'West' | 'East';
    export const HeadingStatusEnum = {
        North: 'North' as HeadingStatusEnum,
        South: 'South' as HeadingStatusEnum,
        West: 'West' as HeadingStatusEnum,
        East: 'East' as HeadingStatusEnum
    };
}